Refactor: various: Replace guint with unsigned int
The GLib documentation says the standard unsigned int should be
preferred in new code. The two are equivalent (via typedef). This means
it's also okay to change public API signatures that use guint.

https://docs.gtk.org/glib/types.html#guint

+++ lib/pacemaker/pcmk_simulate.c
@@ -70,7 +70,7 @@ create_action_name(const pcmk_action_t *action, bool verbose)
 
     if (history_id != NULL) {
         char *key = NULL;
-        guint interval_ms = 0;
+        unsigned int interval_ms = 0;
 
         if (pcmk__guint_from_hash(action->meta, PCMK_META_INTERVAL, 0,
                                   &interval_ms) != pcmk_rc_ok) {
@@ -618,8 +618,7 @@ simulate_resource_action(pcmk__graph_t *graph, pcmk__graph_action_t *action)
     // Certain actions need to be displayed but don't need history entries
     if (pcmk__strcase_any_of(operation, PCMK_ACTION_DELETE,
                              PCMK_ACTION_META_DATA, NULL)) {
-        out->message(out, "inject-rsc-action", resource, operation, node,
-                     (guint) 0);
+        out->message(out, "inject-rsc-action", resource, operation, node, 0);
         goto done; // Confirm action and update graph
     }
